Déplace les tests des routes `/api/*`
