+++ test/routes/testeurMSS.js
@@ -8,9 +8,16 @@ const Referentiel = require('../../src/referentiel');
 const middleware = require('../mocks/middleware');
 
 const testeurMss = () => {
+  let adaptateurMail;
   let depotDonnees;
   let referentiel;
   let serveur;
+
+  const verifieJetonDepose = (reponse, suite) => {
+    const valeurHeader = reponse.headers['set-cookie'][0];
+    expect(valeurHeader).to.match(/^token=.+; path=\/; expires=.+; samesite=strict; httponly$/);
+    suite();
+  };
 
   const verifieRequeteGenereErreurHTTP = (status, messageErreur, requete, suite) => {
     axios(requete)
@@ -24,12 +31,13 @@ const testeurMss = () => {
   };
 
   const initialise = (done) => {
+    adaptateurMail = {};
     middleware.reinitialise();
     referentiel = Referentiel.creeReferentielVide();
     DepotDonnees.creeDepotVide()
       .then((depot) => {
         depotDonnees = depot;
-        serveur = MSS.creeServeur(depotDonnees, middleware, referentiel, {}, false);
+        serveur = MSS.creeServeur(depotDonnees, middleware, referentiel, adaptateurMail, false);
         serveur.ecoute(1234, done);
       });
   };
@@ -37,12 +45,14 @@ const testeurMss = () => {
   const arrete = () => (serveur.arreteEcoute());
 
   return {
+    adaptateurMail: () => adaptateurMail,
     depotDonnees: () => depotDonnees,
     middleware: () => middleware,
     referentiel: () => referentiel,
     arrete,
     initialise,
     verifieRequeteGenereErreurHTTP,
+    verifieJetonDepose,
   };
 };
